Cleaned up file structure and created InfoCard component
