+++ src/App.scss
@@ -87,7 +87,10 @@ body {
     letter-spacing: 0.5px;
 }
 
+// ------------------------------------
 /* NAVBAR */
+// ------------------------------------
+
 .navbar {
     position: absolute;
     bottom: 0;
@@ -169,4 +172,11 @@ body {
             text-align: right;
         }
     }
+}
+
+// ------------------------------------
+// INFOCARD
+// ------------------------------------
+
+.infoCard {
 }
