+++ src/Components/InfoCard.js
@@ -1,0 +1,18 @@
+import React, {useRef, useState } from 'react';
+import ReactDOM from 'react-dom';
+import '../../App.css';
+
+import {Canvas, useLoader, useFrame, useThree} from '@react-three/fiber';
+import { TextureLoader } from 'three/src/loaders/TextureLoader';
+import { OrbitControls, Stars, useTexture } from '@react-three/drei';
+
+
+function InfoCard(props) {
+    return (
+      <div className='infoCard'>
+
+      </div>
+    )
+}
+
+export default InfoCard;
